+++ src/main/kotlin/dtos/product/ProductDetailDtoC.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ProductDetailDtoC (
     val materialId : String,
-    val quantity : Int
+    val quantity : Double
 )
