+++ src/main/kotlin/models/ProductDetailTable.kt
@@ -6,6 +6,6 @@ object ProductDetailTable : Table("products_detail") {
     val id = long("id").autoIncrement()
     val productId = long("product_id").references(ProductTable.id)
     val materialId = long("material_id").references(MaterialTable.id)
-    val quantity = integer("quantity")
+    val quantity = double("quantity")
     override val primaryKey = PrimaryKey(id, name = "pk_product_detail")
 }
